requirements.txt
beautifulsoup4==4.12.3
pdfplumber==0.10.3
python-dotenv==1.0.0
openai==1.12.0
requests==2.31.0
pytz==2023.3.post1
pypdf==3.17.0
pyyaml==6.0.1